Added summary of homework assignment.
Words matter. We are in the business of translating and refining languages, so words matter.

I can't start formulating a solution until I understand the request, so I translated the language of the homework to a language I recognize. The work item language I recognize is presented in this file.

This first translation illustrates the concept of language refinement. The most important factor when translating is to reduce the language to a structured set of "yes/no" questions. Guess what language machines understand?

This exercise ensures comprehension and generates clarifying questions. These clarifications can become a blocker if answers don't come quickly, or if the answer produces more questions. It also exposes any assumptions made before work starts.

In a perfect world, this thorough review of the work item is performed on Spring Planning Monday in order to accurately estimate the necessary sprint tasks. I should already be somewhat familiar with the work item from refinement/sizing sessions.