Created pseudocode solution.
This is a simple, procedural approach to satisfying the Acceptance Criteria in the least elegant way.

This outline helps identify any services that need to be created, and the actions those services will perform. We can even begin to define the inputs and outputs for those actions.

This is the current output of the console app for step 1 of the homework assignment:

   ------------
   App Started
   ------------

      Sort 1 - Started.
      Sort 1 - Complete.

      Sort 2 - Started.
      Sort 2 - Complete.

      Sort 3 - Started.
      Sort 3 - Complete.

   ----------------
   App Completed
   ----------------