Move file model from shared models.
At this point, Homework.ConsoleApp satisfies the acceptance criteria for Homework step 1.

However, RecordRepo.GetRecords() is doing all the work of the File system and the Delimiter system to get the records.

It took some time to refactor the separation of layers and responsibilities of the File, Delimiter, and Record services.